+++ lib/omni_scrapper/scrapper_builder.rb
@@ -1,4 +1,5 @@
 module OmniScrapper
+  # It creates an instance of new Scrapper class with defined configuration.
   class ScrapperBuilder
     attr_accessor :scrapper_name, :config
 
@@ -17,7 +18,7 @@ module OmniScrapper
       crawler = ::OmniScrapper::Crawlers.by_name(config.crawler)
 
       Class.new(crawler) do
-        include OmniScrapper
+        include OmniScrapper::Base
 
         config.anchors.each do |name, options|
           define_method("#{name}_pattern") do
